ECS commented and further refactored, but needs testing before PR
