+++ packages/engine/src/common/functions/hasWindow.ts
@@ -1,0 +1,4 @@
+/**
+ * Detector for browser's "window"
+ */
+export const hasWindow = typeof window !== 'undefined';
